rename sat to ast
