+++ src/lib.rs
@@ -1,11 +1,11 @@
 use once_cell::sync::Lazy;
 
+pub mod ast;
 pub mod bytecode;
 pub mod error;
 pub mod lexer;
 pub mod parser;
 pub mod pest_parser;
-pub mod sat;
 pub mod util;
 pub mod vm;
